🐛 fix(auth) : 중복 닉네임 체크 시리얼라이저, 뷰 수정
